requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
pydantic
python-dotenv
passlib
python-jose[cryptography]
python-multipart
bcrypt